Add files via upload
The purpose of this contribution is to try LDA methodology in Python. LDA is one of many methodologies to infer topic from unstructured text data. Topic modeling is unsupervised learning, we generate many models out of text and afterwards select the one with the highest coherence score. Coherence measures are e.g. Gensim coherence model, TC-W2V. LDA provides high coherence scores compared to NMF or LSA. The interpretability of LDA output is usually not correlated with human brain topic identification.



Bigram - A bigram or digram is a sequence of two adjacent elements from a string of tokens, which are typically letters, syllables, or words. A bigram is an n-gram for n=2.
BOW - Bag of words
Chunksize - controls how many documents are processed at a time in the training algorithm. Increasing chunksize will speed up training, at least as long as the chunk of documents easily fit into memory.
Dirichlet distributions - a way to model random probability mass functions (PMFs) for finite sets.
Dirichlet kernel - it is periodic version of the sinus cardinalis: one the one hand 𝐷𝑛 is obviously periodic (with period 2𝜋. On the other hand, if 𝑛 is large and 𝑥 is small holds:
sin((𝑛+1/2)𝑥)sin(𝑥/2)≈sin((𝑛+1/2)𝑥)𝑥/2=(𝑛+1/2)2sinc((𝑛+1/2)𝑥)
Dump() - dumps() method is used when the objects are required to be in string format and is used for parsing, printing, etc, . The dump() needs the json file name in which the output has to be stored as an argument.
Gibbs sampling - Gibbs sampler is a Markov chain Monte Carlo (MCMC) algorithm for obtaining a sequence of observations which are approximated from a specified multivariate probability distribution, when direct sampling is difficult. Sampler can be blocked or collapsed.
Glob - The glob module is a useful part of the Python standard library. glob (short for global) is used to return all file paths that match a specific pattern.
Latent Dirichlet Allocation - LDA - In natural language processing, Latent Dirichlet Allocation (LDA) is a generative statistical model that explains a set of observations through unobserved groups, and each group explains why some parts of the data are similar.
Lemmatization - the algorithmic process of determining the lemma of a word based on its intended meaning.
TBuckets - measure the quality of LDA, they use singular value decomposition (SVD) to discover important themes in topic words and ILP based optimization to find optimal word-bucket assignments.
TopClus is an unsupervised topic discovery method that jointly models words, documents and topics in a latent spherical space derived from pretrained language model representations.
Topic Coherence - measures score a single topic by measuring the degree of semantic similarity between high scoring words in the topic.
Topic discovery - we model each short document as a Gaussian topic over word embeddings in the vector space.
Topic modeling methodlogies - LDA, LSA (latent semantic analysis), NMF (non negative matrix factorisation). In Scikit-learn, we can generate TF-IDF weighted document term matrix by using TfidfVectorizer.
Triangle - each corner has a topic.
Trigram - a sequence of three adjacent elements from a string of tokens, which are typically letters, syllables, or words. A trigram is an n-gram for n=3.